+++ packages/hash/tests/unit/hash.spec.ts
@@ -2,7 +2,7 @@ import "jest";
 import { Hash } from "../../src";
 
 import { withDOM, asyncWithDOM } from "../../../../tests/utils/dom";
-import { navigateSuite, goSuite } from "../../../../tests/cases";
+import { navigateSuite, goSuite, cancelSuite } from "../../../../tests/cases";
 
 import { TestCase, Suite } from "../../../../tests/types";
 
@@ -158,6 +158,10 @@ describe("Hash constructor", () => {
       });
     });
   });
+});
+
+describe("cancel", () => {
+  runSuite(cancelSuite);
 });
 
 describe("navigate()", () => {