+++ packages/root/src/index.ts
@@ -4,5 +4,12 @@ import locationUtils from "./locationUtils";
 import keyGenerator from "./keyGenerator";
 import prepareNavigate from "./prepareNavigate";
 import navigationConfirmation from "./navigationConfirmation";
+import navigationHandler from "./navigationHandler";
 
-export { locationUtils, keyGenerator, navigationConfirmation, prepareNavigate };
+export {
+  locationUtils,
+  keyGenerator,
+  navigationConfirmation,
+  prepareNavigate,
+  navigationHandler
+};