+++ packages/root/types/index.d.ts
@@ -3,4 +3,5 @@ import locationUtils from "./locationUtils";
 import keyGenerator from "./keyGenerator";
 import prepareNavigate from "./prepareNavigate";
 import navigationConfirmation from "./navigationConfirmation";
-export { locationUtils, keyGenerator, navigationConfirmation, prepareNavigate };
+import navigationHandler from "./navigationHandler";
+export { locationUtils, keyGenerator, navigationConfirmation, prepareNavigate, navigationHandler };