+++ tests/cases/go/cancel-with-replace.ts
@@ -1,5 +1,4 @@
 import "jest";
-import { ignoreFirstCall } from "../../utils/ignoreFirst";
 
 import { AsyncTestCaseArgs } from "../../types";
 
@@ -9,32 +8,44 @@ export default {
   async: true,
   assertions: 1,
   fn: function({ history, resolve }: AsyncTestCaseArgs) {
-    function initialRouter(pending) {
-      pending.finish();
-    }
-    let cancelGo;
-    const goRouter = ignoreFirstCall(function(pending) {
-      cancelGo = pending.cancel;
-      // trigger a replace call and don't resolve the go
-      history.respondWith(replaceRouter);
-      history.navigate("/seven", "replace");
+    let calls = 0;
+    history.respondWith(pending => {
+      switch (calls++) {
+        case 0:
+          pending.finish();
+          history.navigate("/two", "push");
+          break;
+        case 1:
+          pending.finish();
+          history.navigate("/three", "push");
+          break;
+        case 2:
+          pending.finish();
+          history.navigate("/four", "push");
+          break;
+        case 3:
+          pending.finish();
+          history.navigate("/five", "push");
+          break;
+        case 4:
+          pending.finish();
+          history.navigate("/six", "push");
+          break;
+        case 5:
+          pending.finish();
+          history.go(-2);
+          break;
+        case 6:
+          history.navigate("/seven", "replace");
+          break;
+        case 7:
+          pending.finish();
+          expect(history.location).toMatchObject({
+            pathname: "/seven",
+            key: "5.1"
+          });
+          resolve();
+      }
     });
-    const replaceRouter = ignoreFirstCall(function(pending) {
-      cancelGo("replace");
-      setTimeout(() => {
-        expect(history.location.pathname).toBe("/six");
-        resolve();
-      }, 50);
-    });
-
-    history.respondWith(initialRouter);
-    history.navigate("/two", "push"); // 1.0
-    history.navigate("/three", "push"); // 2.0
-    history.navigate("/four", "push"); // 3.0
-    history.navigate("/five", "push"); // 4.0
-    history.navigate("/six", "push"); // 5.0
-
-    history.respondWith(goRouter);
-    history.go(-2);
   }
 };