+++ tests/cases/navigate/finish-not-called.ts
@@ -4,7 +4,7 @@ export default {
   msg: "does nothing if pending.finish() is not called",
   fn: function({ history }: TestCaseArgs) {
     function router() {}
-    history.respondWith(router); // calls router
+    history.respondWith(router);
     history.navigate("/two");
     expect(history.location.pathname).toBe("/one");
   }